feature(minesweeper): GameBoard 의 셀을 여는 책임을 GameBoardCellOpener 에 위임
