+++ src/main/kotlin/minesweeper/domain/GameBoard.kt
@@ -3,8 +3,6 @@ package minesweeper.domain
 import minesweeper.domain.cell.Cell
 import minesweeper.domain.cell.ClosedCell
 import minesweeper.domain.cell.Location
-import minesweeper.domain.cell.NumberOfAdjacentMines
-import minesweeper.domain.strategy.AdjacentLocationDirection
 
 class GameBoard private constructor(
     val area: Area,
@@ -21,51 +19,6 @@ class GameBoard private constructor(
     fun openAll(): GameBoard {
         val allOpenedCells = cells.map { if (it is ClosedCell) it.open() else it }
         return GameBoard(area, Cells(allOpenedCells))
-    }
-
-    fun open(location: Location): GameBoard {
-        val foundedCell = this.find(location) ?: throw IllegalArgumentException("해당 위치를 가진 셀이 존재하지 않습니다")
-        if (foundedCell !is ClosedCell) {
-            return this
-        }
-        if (foundedCell.hasLandmine) {
-            val cellsWithOpenCells = cells.map { if (it == foundedCell && it is ClosedCell) it.open() else it }
-            return GameBoard(area, Cells(cellsWithOpenCells))
-        }
-        if (foundedCell.numberOfAdjacentLandmines > NumberOfAdjacentMines.ZERO) {
-            val cellsWithOpenCells = cells.map { if (it == foundedCell && it is ClosedCell) it.open() else it }
-            return GameBoard(area, Cells(cellsWithOpenCells))
-        }
-
-        val candidates = mutableListOf<Cell>()
-        val queue = ArrayDeque<Cell>()
-        val visitedCells = mutableListOf<Cell>()
-
-        visitedCells.add(foundedCell)
-        queue.add(foundedCell)
-        candidates.add(foundedCell)
-
-        while (queue.isNotEmpty()) {
-            val nextCell = queue.removeFirst()
-            val allAdjacentLocations =
-                AdjacentLocationDirection.allAdjacentLocations(nextCell.location)
-            allAdjacentLocations.forEach { adjacentLocation ->
-                val adjacentCell = find(adjacentLocation)
-                adjacentCell?.let { cell ->
-                    if (cell is ClosedCell && cell !in visitedCells && !cell.hasLandmine) {
-                        visitedCells.add(cell)
-                        candidates.add(cell)
-
-                        if (cell.numberOfAdjacentLandmines == NumberOfAdjacentMines.ZERO) {
-                            queue.add(cell)
-                        }
-                    }
-                }
-            }
-        }
-
-        val cellsWithOpenCells = cells.map { if (candidates.contains(it) && it is ClosedCell) it.open() else it }
-        return GameBoard(area, Cells(cellsWithOpenCells))
     }
 
     companion object {
